Deburgging
Still stuck with the cmake config. So we just move methods implementation from
.cpp to .hpp the time to fix this issue.

Also notice when using reverse loop with the value of coarsest_level(class member) we end in an infinite loop. So we impose coarsest_value to be 0 for now. This should also be fixed.

The FCycle and WCycle should be implement in multigrid_base.hpp